requirements.txt
# Python Trading Algorithm Dependencies for Fundezy Platform
# Install with: pip install -r requirements.txt

# Core dependencies
requests>=2.28.0
pandas>=1.5.0
numpy>=1.24.0

# Optional dependencies for advanced strategies
# Uncomment as needed:

# Technical analysis
# ta-lib>=0.4.25
# talib>=0.4.25

# Machine learning
# scikit-learn>=1.2.0
# tensorflow>=2.12.0
# torch>=2.0.0

# Data visualization
# matplotlib>=3.6.0
# plotly>=5.14.0

# Environment management
python-dotenv>=1.0.0

# Development tools
# pytest>=7.2.0
# black>=23.0.0
# flake8>=6.0.0